kernel/os: Fix integer overlow on tickless calculations
When using large OS_TICKS_PER_SEC value there will be integer overflow
error on multiplication. We can just explicitly perform calculations
using uint64_t, this done at build time anyway and the result will be
uint32_t unless some really big values are used in which case we do not
care. This is to allow CMAC work at 31250 ticks per sec.

+++ kernel/os/src/os.c
@@ -79,8 +79,8 @@ static struct hal_timer os_wdog_monitor;
  */
 int g_os_started;
 
-#define MIN_IDLE_TICKS  (MYNEWT_VAL(OS_IDLE_TICKLESS_MS_MIN) * OS_TICKS_PER_SEC / 1000)
-#define MAX_IDLE_TICKS  (MYNEWT_VAL(OS_IDLE_TICKLESS_MS_MAX) * OS_TICKS_PER_SEC / 1000)
+#define MIN_IDLE_TICKS  (uint32_t)((uint64_t)MYNEWT_VAL(OS_IDLE_TICKLESS_MS_MIN) * OS_TICKS_PER_SEC / 1000)
+#define MAX_IDLE_TICKS  (uint32_t)((uint64_t)MYNEWT_VAL(OS_IDLE_TICKLESS_MS_MAX) * OS_TICKS_PER_SEC / 1000)
 
 /**
  * Idle operating system task, runs when no other tasks are running.
